docflo/api_service/requirements.txt
gunicorn[gthread]
cryptography
mysql-connector-python
flask
peewee
toml
google-api-python-client
pytest
requests
apscheduler
pymysql